ajuste en modelo Expediente

+++ db/migrate/20220725182322_create_expedients.rb
@@ -5,9 +5,9 @@ class CreateExpedients < ActiveRecord::Migration[7.0]
       t.string :apellido_solicitante
       t.string :tipo_persona
       t.string :edad
-      t.integer :expediente_status_id
+      t.bigint :expediente_status_id
       t.date :fecha_nacimiento
-      t.integer :tipo_solicitud_id
+      t.bigint :tipo_solicitud_id
       t.integer :cantidad_solicitada
       t.string :identificador_unico
       t.string :email
